rename files and remove nesting
