+++ libs/design/src/atoms/form/native-select/README.md
@@ -1,4 +1,4 @@
-# Select Component
+# Native Select Component
 
 `daff-native-select` is a form control element that can be used in forms to select a value from a set of options.
