+++ libs/design/src/atoms/form/native-select/public_api.ts
@@ -1,2 +1,2 @@
-export * from './select/select.component';
-export { DaffNativeSelectModule } from './select.module';
+export * from './native-select.component';
+export { DaffNativeSelectModule } from './native-select.module';
